Added client sided join & quit handlers
Players can now actually exist in the world and other players also join and leave the scene via the server!

+++ Assets/Networking/utils/PacketUtils.cs
@@ -17,7 +17,7 @@ public abstract class PacketUtils {
 
     public static float FromPacketString(string packetNumber) {
         int integerFromPacket = int.Parse(packetNumber);
-        float number = integerFromPacket / packetFloatToIntMultiplier;
+        float number = (float)integerFromPacket / packetFloatToIntMultiplier;
         return number;
     }
